#397 | Move note card and listing component to appropriate package
